+++ frontend/app/experiments/(experiment)/(forms)/DynamicArrayField.tsx
@@ -11,7 +11,6 @@ import { FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessa
 import { Input } from "@/components/ui/input";
 import { LabelDescription } from "@/components/ui/label";
 import { Textarea } from "@/components/ui/textarea";
-import { cn } from "@/lib/utils";
 
 interface DynamicFieldProps {
   form: UseFormReturn<any>;
